backend/django/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django>=5.2.1
djangorestframework>=3.16.0
h11==0.14.0
psycopg2-binary>=2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv>=1.0.0
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.28.0
